+++ src/BugsnagUnity/Native/Cocoa/NativeClient.cs
@@ -9,26 +9,38 @@ namespace BugsnagUnity
   class NativeClient : INativeClient
   {
     public IConfiguration Configuration { get; }
-
     public IBreadcrumbs Breadcrumbs { get; }
-
     public IDelivery Delivery { get; }
 
     IntPtr NativeConfiguration { get; }
 
-    NativeClient(IConfiguration configuration, IntPtr nativeConfiguration, IBreadcrumbs breadcrumbs)
+    public NativeClient(IConfiguration configuration)
     {
       Configuration = configuration;
-      NativeConfiguration = nativeConfiguration;
-
+      NativeConfiguration = CreateNativeConfig(configuration);
       NativeCode.bugsnag_startBugsnagWithConfiguration(NativeConfiguration, NotifierInfo.NotifierVersion);
-
       Delivery = new Delivery();
-      Breadcrumbs = breadcrumbs;
+      Breadcrumbs = new Breadcrumbs(NativeConfiguration);
     }
 
-    public NativeClient(Configuration configuration) : this(configuration, configuration.NativeConfiguration, new Breadcrumbs(configuration))
-    {
+    /**
+     * Transforms an IConfiguration C# object into a Cocoa Configuration object.
+     */
+    IntPtr CreateNativeConfig(IConfiguration config) {
+      IntPtr obj = NativeCode.bugsnag_createConfiguration(config.ApiKey);
+      NativeCode.bugsnag_setAutoNotify(obj, config.AutoNotify);
+      NativeCode.bugsnag_setReleaseStage(obj, config.ReleaseStage);
+      NativeCode.bugsnag_setAppVersion(obj, config.AppVersion);
+      NativeCode.bugsnag_setNotifyUrl(obj, config.Endpoint.ToString());
+
+      if (config.Context != null) {
+        NativeCode.bugsnag_setContext(obj, config.Context);
+      }
+      var releaseStages = config.NotifyReleaseStages;
+      if (releaseStages != null) {
+        NativeCode.bugsnag_setNotifyReleaseStages(obj, releaseStages, releaseStages.Length);
+      }
+      return obj;
     }
 
     public void PopulateApp(App app)
@@ -156,6 +168,16 @@ namespace BugsnagUnity
     public void SetUser(User user)
     {
       NativeCode.bugsnag_setUser(user.Id, user.Name, user.Email);
+    }
+
+    public void SetContext(string context)
+    {
+      NativeCode.bugsnag_setContext(NativeConfiguration, context);
+    }
+
+    public void SetAutoNotify(bool autoNotify)
+    {
+      NativeCode.bugsnag_setAutoNotify(NativeConfiguration, autoNotify);
     }
   }
 }