+++ src/BugsnagUnity/Native/Windows/NativeClient.cs
@@ -12,7 +12,7 @@ namespace BugsnagUnity
 
     public IDelivery Delivery { get; }
 
-    public NativeClient(Configuration configuration)
+    public NativeClient(IConfiguration configuration)
     {
       Configuration = configuration;
       Breadcrumbs = new Breadcrumbs(configuration);
@@ -65,6 +65,12 @@ namespace BugsnagUnity
     }
 
     public void SetUser(User user)
+    {
+    }
+    public void SetContext(string context)
+    {
+    }
+    public void SetAutoNotify(bool autoNotify)
     {
     }