+++ test/mobile/features/fixtures/nativeplugin/ios/BugsnagNative/BugsnagNative.h
@@ -1,0 +1,15 @@
+//
+//  BugsnagNative.h
+//  BugsnagNative
+//
+//  Created by Alexander Moinet on 14/09/2018.
+//  Copyright © 2018 Bugsnag. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface BugsnagNative : NSObject
+
++ (void) crashMe;
+
+@end